refactored useable items as projectiles
