+++ LoZGame/LoZGame.cs
@@ -16,7 +16,7 @@ namespace LoZClone
         private IController keyboardController;
         private ItemManager itemManager;
         private BlockManager blockManager;
-        private InventoryManager inventoryManager;
+        private ProjectileManager projectileManager;
 
         public LoZGame()
         {
@@ -29,8 +29,8 @@ namespace LoZClone
             link = new Link(this);
             itemManager = new ItemManager();
             blockManager = new BlockManager();
-            inventoryManager = new InventoryManager();
-            commandLoader = new CommandLoader(this, link, itemManager, blockManager, inventoryManager);
+            projectileManager = new ProjectileManager();
+            commandLoader = new CommandLoader(this, link, itemManager, blockManager, projectileManager);
             keyboardController = new KeyboardController(this, commandLoader);
             base.Initialize();
         }
@@ -40,8 +40,9 @@ namespace LoZClone
             LinkSpriteFactory.Instance.LoadAllTextures(Content);
             ItemSpriteFactory.Instance.LoadAllTextures(Content);
             BlockSpriteFactory.Instance.LoadAllTextures(Content);
-            itemManager.loadSprites(120, 120);
-            blockManager.loadSprites(240, 240);
+            ProjectileSpriteFactory.Instance.LoadAllTextures(Content);
+            itemManager.loadSprites(300, 240);
+            blockManager.loadSprites(500, 240);
         }
         protected override void UnloadContent()
         {
@@ -52,7 +53,7 @@ namespace LoZClone
             link.Update();
             itemManager.currentItem.Update();
             blockManager.currentBlock.Update();
-            inventoryManager.Update();
+            projectileManager.Update();
             base.Update(gameTime);
         }
         protected override void Draw(GameTime gameTime)
@@ -62,7 +63,7 @@ namespace LoZClone
             link.Draw();
             itemManager.currentItem.Draw(spriteBatch);
             blockManager.currentBlock.Draw(spriteBatch, new Vector2(240, 150), Color.White);
-            inventoryManager.Draw(spriteBatch);
+            projectileManager.Draw(spriteBatch);
             spriteBatch.End();
             base.Draw(gameTime);
         }
