+++ LoZGame/controller/Commands/CommandTwo.cs
@@ -3,15 +3,15 @@
     public class CommandTwo : ICommand
     {
         IPlayer player;
-        InventoryManager inventory;
-        public CommandTwo(IPlayer player, InventoryManager inventory)
+        ProjectileManager projectile;
+        public CommandTwo(IPlayer player, ProjectileManager projectile)
         {
             this.player = player;
-            this.inventory = inventory;
+            this.projectile = projectile;
         }
         public void execute()
         {   
-            inventory.addItem(inventory.Arrow, ((Link)player).CurrentLocation, ((Link)player).CurrentDirection);
+            projectile.addItem(projectile.Arrow, ((Link)player).CurrentLocation, ((Link)player).CurrentDirection);
         }
     }
 }
