+++ LoZGame/interfaces/IProjectile.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework;
+
+namespace LoZClone
+{
+    interface IProjectile
+    {
+        bool IsExpired { get; }
+
+        int Instance { get; }
+
+        Vector2 location { get; }
+
+        void Update();
+
+        void Draw(SpriteBatch spriteBatch);
+
+    }
+}
