--- LoZGame/interfaces/IUsableItem.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LoZClone
-{
-    interface IUsableItem : IItemSprite
-    {
-        bool IsExpired { get; }
-
-        int Instance { get; }
-
-    }
-}
