+++ LoZGame/sprites/ItemSpriteClasses/RedCandle.cs
@@ -8,25 +8,12 @@ using Microsoft.Xna.Framework.Graphics;
 
 namespace LoZClone
 {
-    class RedCandle : IItemSprite, IUsableItem
+    class RedCandle : IItemSprite
     {
-        private Texture2D Texture;      // the texture to pull frames from
-        private Rectangle firstFrame;
-        private Rectangle secondFrame;
+        private Texture2D Texture;
         private Rectangle currentFrame;
         private int scale;
-        private string direction;
-        private Vector2 destination;
         private int lifeTime;
-        private int distTravelled;
-
-        private static int lifeTimeMax = 210;
-        private static int travelDistance = 256;
-        private static int frameDelay = 10;
-
-        private int instance;
-        private bool expired;
-        private bool moving;
 
 
         public Vector2 location { get; set; }
@@ -36,90 +23,12 @@ namespace LoZClone
             currentFrame = new Rectangle(160, 0, 6, 16);
             location = loc;
             this.scale = scale;
-            moving = false;
-            lifeTime = lifeTimeMax;
-        }
-
-        public RedCandle(Texture2D texture, Vector2 loc, string direction, int scale, int instance)
-        {
-            lifeTime = lifeTimeMax;
-            Texture = texture;
-            firstFrame = new Rectangle(0, 0, 20, 20);
-            secondFrame = new Rectangle(0, 30, 20, 20);
-            currentFrame = firstFrame;
-            this.scale = scale;
-            this.instance = instance;
-            expired = false;
-            this.direction = direction;
-            moving = true;
-            distTravelled = 1;
-
-            if (direction.Equals("Up"))
-            {
-                location = new Vector2(loc.X + 13, loc.Y);
-                destination = new Vector2(this.location.X, this.location.Y - travelDistance);
-            }
-            else if (direction.Equals("Left"))
-            {
-                location = new Vector2(loc.X, loc.Y + 8);
-                destination = new Vector2(this.location.X - travelDistance, this.location.Y);
-            }
-            else if (direction.Equals("Right"))
-            {
-                location = new Vector2(loc.X + 32, loc.Y + 8);
-                destination = new Vector2(this.location.X + travelDistance, this.location.Y);
-            }
-            else
-            {
-                location = new Vector2(loc.X + 13, loc.Y + 32);
-                destination = new Vector2(this.location.X, this.location.Y + travelDistance);
-            }
-        }
-
-        public bool IsExpired
-        {
-            get { return expired; }
-        }
-
-        public int Instance
-        {
-            get { return instance; }
-        }
-
-        private void nextFrame()
-        {
-            if (currentFrame == firstFrame)
-            {
-                currentFrame = secondFrame;
-            }
-            else
-            {
-                currentFrame = firstFrame;
-            }
+            lifeTime = 0;
         }
 
         public void Update()
         {
-            if (moving)
-            {
-                if (lifeTime % frameDelay == 0)
-                {
-                    this.nextFrame();
-                }
-                if (lifeTime >= lifeTimeMax / 3)
-                {
-                    float xdiff = (destination.X - location.X) / 8;
-                    float ydiff = (destination.Y - location.Y) / 8;
-                    float denom = 2 * distTravelled + 1 / distTravelled;
-                    this.location = new Vector2(this.location.X + (xdiff / denom), this.location.Y + (ydiff / denom));
-                }
-                else if (lifeTime <= 0)
-                {
-                    expired = true;
-                }
-                distTravelled++;
-                lifeTime--;
-            }
+            lifeTime++;
         }
 
         public void Draw(SpriteBatch spriteBatch)
