+++ LoZGame/sprites/ProjectileSpriteClasses/TriforceProjectile.cs
@@ -6,9 +6,10 @@ using System.Threading.Tasks;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
+
 namespace LoZClone
 {
-    class TriForce : IItemSprite
+    class TriforceProjectile : IProjectile
     {
         private Texture2D Texture;      // the texture to pull frames from
         private Rectangle currentFrame;
@@ -16,20 +17,24 @@ namespace LoZClone
         private Rectangle secondFrame;
         private int lifeTime;
         private int scale;
-        private float rotation;
+        private int instance;
+        private bool expired;
+        private bool isStatic;
         public Vector2 location { get; set; }
-        public TriForce(Texture2D texture, Vector2 loc, int scale)
+
+        public TriforceProjectile(Texture2D texture, Vector2 loc, int scale, int instance)
         {
             Texture = texture;
             firstFrame = new Rectangle(275, 0, 10, 16);
             secondFrame = new Rectangle(275, 16, 10, 16);
             currentFrame = firstFrame;
-            lifeTime = 0;
-            location = loc;
-            this.rotation = 0;
+            lifeTime = 200;
+            location = new Vector2(loc.X + 11, loc.Y - 32);
             this.scale = scale;
+            expired = false;
+            this.instance = instance;
+            isStatic = false;
         }
-
         private void nextFrame()
         {
             if (currentFrame == firstFrame)
@@ -41,22 +46,45 @@ namespace LoZClone
                 currentFrame = firstFrame;
             }
         }
-        
+        public bool IsExpired
+        {
+            get { return expired; }
+        }
+
+        public int Instance
+        {
+            get { return instance; }
+        }
+
         public void Update()
         {
-            lifeTime++;
+            lifeTime--;
+
+            if (isStatic)
+            {
+                if (lifeTime <= 0)
+                {
+                    lifeTime = 200;
+                }
+            }
+            else
+            {
+                if (lifeTime <= 0)
+                {
+                    expired = true;
+                }
+            }
 
             if (lifeTime % 4 == 0)
             {
                 this.nextFrame();
-                lifeTime = 0;
             }
         }
 
         public void Draw(SpriteBatch spriteBatch)
         {
-            spriteBatch.Draw(Texture, this.location, currentFrame, Color.White, rotation, new Vector2(2, 8), scale, SpriteEffects.None, 0f);
+            Rectangle dest = new Rectangle((int)location.X, (int)location.Y, firstFrame.Width * scale, firstFrame.Height * scale);
+            spriteBatch.Draw(Texture, dest, currentFrame, Color.White);
         }
-
     }
 }
