+++ LoZGame/util/ItemSpriteFactory.cs
@@ -100,10 +100,6 @@ namespace LoZClone
         {
             return new TriForce(itemSpriteSheet, loc, scale);
         }
-        public IItemSprite Triforce(Vector2 loc, int scale, int instance)
-        {
-            return new TriForce(itemSpriteSheet, loc, scale, instance);
-        }
         public IItemSprite YellowRupee(Vector2 loc, int scale)
         {
             return new YellowRupee(itemSpriteSheet, loc, scale);
@@ -172,25 +168,13 @@ namespace LoZClone
         {
             return new Boomerang(itemSpriteSheet, loc, scale);
         }
-        public IItemSprite Boomerang(Link player, int scale, int instance)
-        {
-            return new Boomerang(itemSpriteSheet, player, scale, instance);
-        }
         public IItemSprite MagicBoomerang(Vector2 loc, int scale)
         {
             return new MagicBoomerang(itemSpriteSheet, loc, scale);
         }
-        public IItemSprite MagicBoomerang(Link player, int scale, int instance)
-        {
-            return new MagicBoomerang(itemSpriteSheet, player, scale, instance);
-        }
         public IItemSprite Bomb(Vector2 loc, int scale)
         {
             return new Bomb(itemSpriteSheet, loc, scale);
-        }
-        public IItemSprite Bomb(Vector2 loc, string direction, int scale, int instance)
-        {
-            return new Bomb(itemSpriteSheet, loc, direction, scale, instance);
         }
         public IItemSprite Bow(Vector2 loc, int scale)
         {
@@ -200,33 +184,17 @@ namespace LoZClone
         {
             return new Arrow(itemSpriteSheet, loc, scale);
         }
-        public IItemSprite Arrow(Vector2 loc, string direction, int scale, int instance)
-        {
-            return new Arrow(itemSpriteSheet, loc, direction, scale, instance);
-        }
         public IItemSprite SilverArrow(Vector2 loc, int scale)
         {
             return new SilverArrow(itemSpriteSheet, loc, scale);
-        }
-        public IItemSprite SilverArrow(Vector2 loc, string direction, int scale, int instance)
-        {
-            return new SilverArrow(itemSpriteSheet, loc, direction, scale, instance);
         }
         public IItemSprite RedCandle(Vector2 loc, int scale)
         {
             return new RedCandle(itemSpriteSheet, loc, scale);
         }
-        public IItemSprite RedCandle(Vector2 loc, string direction, int scale, int instance)
-        {
-            return new RedCandle(fireSpriteSheet, loc, direction, scale, instance);
-        }
         public IItemSprite BlueCandle(Vector2 loc, int scale)
         {
             return new BlueCandle(itemSpriteSheet, loc, scale);
-        }
-        public IItemSprite BlueCandle(Vector2 loc, string direction, int scale, int instance)
-        {
-            return new BlueCandle(fireSpriteSheet, loc, direction, scale, instance);
         }
         public IItemSprite RedRing(Vector2 loc, int scale)
         {
